backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
groq==0.11.0
httpx==0.27.0
supabase==2.7.0
networkx==3.3
celery==5.4.0
redis==5.1.0
langchain==0.3.0
langchain-community==0.3.0
peft==0.13.0
trl==0.11.0
transformers==4.45.0
bitsandbytes==0.44.0
arxiv==2.1.0
youtube-transcript-api==0.6.2
trafilatura==1.12.0
beautifulsoup4==4.12.3
PyGithub==2.4.0
PyMuPDF==1.24.0
pdfplumber==0.11.0
biopython==1.84
numpy==1.26.4
pydantic==2.9.0
pydantic-settings==2.5.0
python-multipart==0.0.9
sse-starlette==2.1.0
python-jose==3.3.0
passlib==1.7.4
resend==2.4.0
huggingface_hub==0.25.0
datasets==3.0.0
Pillow==10.4.0
reportlab==4.2.0
sentence-transformers==3.1.0
